# event-manager/src/main/resources/static/event-manager-openapi.yaml
...
openapi: 3.0.0
info:
  description: API for managing event locations
  title: Event Management Service
  version: 1.0.0
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
...
paths:
...
  "/events":
    post:
      description: |-
        Учтите валидацию всех полей. Все поля обязательные при создании, также должна быть соблюдена логика (например время не может быть в прошлом, стомость > 0 и т.д.) Локация должна существовать и ее вместимость должна позволять провести это мероприятие, т.е. кол-во мест должно быть достаточно для всех участников. Организатор (создатель) мероприятия не учитывается как участник.

        Изменения по итерациям:
        - Итерация 3: endpoint добавлен.
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/EventCreateRequestDto"
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/EventDto"
          description: Мероприятие успешно создано. Возвращается созданное событие.
        '400':
          content:
            application/json:
              example:
                message: Некорректный запрос
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: '2024-12-03T10:15:30'
              schema:
                "$ref": "#/components/schemas/ErrorMessageResponse"
          description: Запрос с невалидными данными. Например, попытка занять занятый
            логин или некорректные параметры запроса.
        '401':
          content:
            application/json:
              example:
                message: Необходима аутентификация
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: '2024-12-03T10:15:30'
              schema:
                "$ref": "#/components/schemas/ErrorMessageResponse"
          description: Аутентификация отсутствует или не удалась
        '403':
          content:
            application/json:
              example:
                message: Недостаточно прав для выполнения операции
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: '2024-12-03T10:15:30'
              schema:
                "$ref": "#/components/schemas/ErrorMessageResponse"
          description: У пользователя недостаточно прав для выполнения этой операции
        '404':
          content:
            application/json:
              example:
                message: Сущность не найдена
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: '2024-12-03T10:15:30'
              schema:
                "$ref": "#/components/schemas/ErrorMessageResponse"
          description: Сущность не найдена.
        '500':
          content:
            application/json:
              example:
                message: Внутренняя ошибка сервера
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: '2024-12-03T10:15:30'
              schema:
                "$ref": "#/components/schemas/ErrorMessageResponse"
          description: Внутренняя ошибка сервера
      summary: Создание нового мероприятия. Allowed roles=[USER]
      tags:
      - Events
  "/events/{eventId}":
    delete:
      description: |-
        Доступно только ADMIN, либо создателю мероприятия. Т.е. роль из JWT токена должна быть ADMIN, либо userId должен быть равен создателю меропрития. Должно быть реализовано "мягкое(soft) удаление" - на самом деле строка в БД не удалется, только меняется статус мероприятия на CANCELLED. При этом нужно учесть то, что не каждое мероприятие можно отменить (можно отменить только те мероприятия, которые еще не начались).

        Изменения по итерациям:
        - Итерация 3: endpoint добавлен.
        - Итерация 4: при отмене события добавлена публикация в Kafka доменного события по общему контракту итерации 4: `messageId`, `eventType`, `eventId`, `occurredAt`, `ownerId`, `changedById`, `subscribers`, `changes[]`; отмена отражается как изменение поля `status`.
        - Итерация 5: изменена внутренняя логика после отмены события: добавлена инвалидация кеша события по id.
      parameters:
      - description: Уникальный идентификатор мероприятия
        in: path
        name: eventId
        required: true
        schema:
          example: 10
          format: int64
          type: integer
      responses:
        '204':
          description: Мероприятие успешно удалено
        '400':
          content:
            application/json:
              example:
                message: Некорректный запрос
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: '2024-12-03T10:15:30'
              schema:
                "$ref": "#/components/schemas/ErrorMessageResponse"
          description: Запрос с невалидными данными. Например, попытка занять занятый
            логин или некорректные параметры запроса.
        '401':
          content:
            application/json:
              example:
                message: Необходима аутентификация
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: '2024-12-03T10:15:30'
              schema:
                "$ref": "#/components/schemas/ErrorMessageResponse"
          description: Аутентификация отсутствует или не удалась
        '403':
          content:
            application/json:
              example:
                message: Недостаточно прав для выполнения операции
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: '2024-12-03T10:15:30'
              schema:
                "$ref": "#/components/schemas/ErrorMessageResponse"
          description: У пользователя недостаточно прав для выполнения этой операции
        '404':
          content:
            application/json:
              example:
                message: Сущность не найдена
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: '2024-12-03T10:15:30'
              schema:
                "$ref": "#/components/schemas/ErrorMessageResponse"
          description: Сущность не найдена.
        '500':
          content:
            application/json:
              example:
                message: Внутренняя ошибка сервера
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: '2024-12-03T10:15:30'
              schema:
                "$ref": "#/components/schemas/ErrorMessageResponse"
          description: Внутренняя ошибка сервера
      summary: Удаление мероприятия. Allowed roles=[USER, ADMIN]
      tags:
      - Events
    get:
      description: |-
        Предусмотрите обработку ошибок, если мероприятие не найдено.

        Изменения по итерациям:
        - Итерация 3: endpoint добавлен.
        - Итерация 5: изменена внутренняя логика чтения: используется cache-aside для события по id с Redis; кеш также инвалидируется при смене статуса события scheduler-ом.
      parameters:
      - description: Уникальный идентификатор мероприятия
        in: path
        name: eventId
        required: true
        schema:
          example: 10
          format: int64
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/EventDto"
          description: Данные мероприятия
        '400':
          content:
            application/json:
              example:
                message: Некорректный запрос
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: '2024-12-03T10:15:30'
              schema:
                "$ref": "#/components/schemas/ErrorMessageResponse"
          description: Запрос с невалидными данными. Например, попытка занять занятый
            логин или некорректные параметры запроса.
        '401':
          content:
            application/json:
              example:
                message: Необходима аутентификация
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: '2024-12-03T10:15:30'
              schema:
                "$ref": "#/components/schemas/ErrorMessageResponse"
          description: Аутентификация отсутствует или не удалась
        '403':
          content:
            application/json:
              example:
                message: Недостаточно прав для выполнения операции
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: '2024-12-03T10:15:30'
              schema:
                "$ref": "#/components/schemas/ErrorMessageResponse"
          description: У пользователя недостаточно прав для выполнения этой операции
        '404':
          content:
            application/json:
              example:
                message: Сущность не найдена
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: '2024-12-03T10:15:30'
              schema:
                "$ref": "#/components/schemas/ErrorMessageResponse"
          description: Сущность не найдена.
        '500':
          content:
            application/json:
              example:
                message: Внутренняя ошибка сервера
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: '2024-12-03T10:15:30'
              schema:
                "$ref": "#/components/schemas/ErrorMessageResponse"
          description: Внутренняя ошибка сервера
      summary: Получение мероприятия по ID. Allowed roles=[USER, ADMIN]
      tags:
      - Events
    put:
      description: |-
        Доступно только ADMIN, либо создателю мероприятия. Т.е. роль из JWT токена должна быть ADMIN, либо userId должен быть равен создателю меропрития. Учтите то, как можно менять мероприятие. Валидируйте входные значения, например maxPlaces должно быть больше, чем уже записанных пользователей, стоимость > 0, длительность > 0 и т.д.

        Изменения по итерациям:
        - Итерация 3: endpoint добавлен.
        - Итерация 4: при изменении события добавлена публикация в Kafka доменного события по общему контракту итерации 4: `messageId`, `eventType`, `eventId`, `occurredAt`, `ownerId`, `changedById`, `subscribers`, `changes[]`. Если изменение инициировано не пользователем, `changedById` может быть `null`.
        - Итерация 5: изменена внутренняя логика после обновления: добавлена инвалидация кеша события по id.
      parameters:
      - description: Уникальный идентификатор мероприятия
        in: path
        name: eventId
        required: true
        schema:
          example: 10
          format: int64
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/EventUpdateRequestDto"
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/EventDto"
          description: Обновленное мероприятие
        '400':
          content:
            application/json:
              example:
                message: Некорректный запрос
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: '2024-12-03T10:15:30'
              schema:
                "$ref": "#/components/schemas/ErrorMessageResponse"
          description: Запрос с невалидными данными. Например, попытка занять занятый
            логин или некорректные параметры запроса.
        '401':
          content:
            application/json:
              example:
                message: Необходима аутентификация
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: '2024-12-03T10:15:30'
              schema:
                "$ref": "#/components/schemas/ErrorMessageResponse"
          description: Аутентификация отсутствует или не удалась
        '403':
          content:
            application/json:
              example:
                message: Недостаточно прав для выполнения операции
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: '2024-12-03T10:15:30'
              schema:
                "$ref": "#/components/schemas/ErrorMessageResponse"
          description: У пользователя недостаточно прав для выполнения этой операции
        '404':
          content:
            application/json:
              example:
                message: Сущность не найдена
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: '2024-12-03T10:15:30'
              schema:
                "$ref": "#/components/schemas/ErrorMessageResponse"
          description: Сущность не найдена.
        '500':
          content:
            application/json:
              example:
                message: Внутренняя ошибка сервера
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: '2024-12-03T10:15:30'
              schema:
                "$ref": "#/components/schemas/ErrorMessageResponse"
          description: Внутренняя ошибка сервера
      summary: Обновление мероприятия. Allowed roles=[USER, ADMIN]
      tags:
      - Events
  "/events/search":
    post:
      description: |-
        Поиск мероприятий по фильтру. Если все поля пустые, должен вернуться список всех мероприятий.
        Поддержите весь набор фильтров из `EventSearchRequestDto`: имя, диапазоны по местам,
        дате старта, стоимости, длительности, а также `locationId` и `eventStatus`.
        Для поля `name` используйте точное совпадение, а не поиск по подстроке.

        Изменения по итерациям:
        - Итерация 3: endpoint добавлен.
      parameters:
        - name: pageNum
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
            minimum: 0
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 2
            maximum: 20
            minimum: 2
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/EventSearchRequestDto"
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  "$ref": "#/components/schemas/EventDto"
                type: array
          description: Список мероприятий, соответствующих заданным фильтрам
        '400':
          content:
            application/json:
              example:
                message: Некорректный запрос
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: '2024-12-03T10:15:30'
              schema:
                "$ref": "#/components/schemas/ErrorMessageResponse"
          description: Запрос с невалидными данными. Например, попытка занять занятый
            логин или некорректные параметры запроса.
        '401':
          content:
            application/json:
              example:
                message: Необходима аутентификация
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: '2024-12-03T10:15:30'
              schema:
                "$ref": "#/components/schemas/ErrorMessageResponse"
          description: Аутентификация отсутствует или не удалась
        '403':
          content:
            application/json:
              example:
                message: Недостаточно прав для выполнения операции
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: '2024-12-03T10:15:30'
              schema:
                "$ref": "#/components/schemas/ErrorMessageResponse"
          description: У пользователя недостаточно прав для выполнения этой операции
        '500':
          content:
            application/json:
              example:
                message: Внутренняя ошибка сервера
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: '2024-12-03T10:15:30'
              schema:
                "$ref": "#/components/schemas/ErrorMessageResponse"
          description: Внутренняя ошибка сервера
      summary: Поиск мероприятий по фильтру. Allowed roles=[USER, ADMIN]
      tags:
      - Events
  "/events/my":
    get:
      description: |-
        Все мероприятия созданные пользователем, который выполняет запрос.

        Изменения по итерациям:
        - Итерация 3: endpoint добавлен.
      parameters:
        - name: pageNum
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
            minimum: 0
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 2
            maximum: 20
            minimum: 2
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  "$ref": "#/components/schemas/EventDto"
                type: array
          description: Список мероприятий, созданных пользователем
        '401':
          content:
            application/json:
              example:
                message: Необходима аутентификация
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: '2024-12-03T10:15:30'
              schema:
                "$ref": "#/components/schemas/ErrorMessageResponse"
          description: Аутентификация отсутствует или не удалась
        '403':
          content:
            application/json:
              example:
                message: Недостаточно прав для выполнения операции
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: '2024-12-03T10:15:30'
              schema:
                "$ref": "#/components/schemas/ErrorMessageResponse"
          description: У пользователя недостаточно прав для выполнения этой операции
        '500':
          content:
            application/json:
              example:
                message: Внутренняя ошибка сервера
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: '2024-12-03T10:15:30'
              schema:
                "$ref": "#/components/schemas/ErrorMessageResponse"
          description: Внутренняя ошибка сервера
      summary: Получение всех мероприятий, созданных текущим пользователем. Allowed
        roles=[USER]
      tags:
      - Events
...
components:
...
  schemas:
    ErrorMessageResponse:
      example:
        message: Сообщение об ошибке
        detailedMessage: Более детальное сообщение об ошибке
        dateTime: '2024-12-03T10:15:30'
      properties:
        message:
          description: Человеко-читаемое сообщение об ошибке
          type: string
        detailedMessage:
          description: Более детальное сообщение об ошибке. Например, содержимое exception
          type: string
        dateTime:
          description: Дата и время возникновения ошибки. Формат "YYYY-MM-DDThh:mm:ss"
          format: date-time
          type: string
      required:
      - dateTime
      - detailedMessage
      - message
      type: object
...
    EventDto:
      example:
        occupiedPlaces: 7
        date: '2000-01-23T04:56:07.000+00:00'
        duration: 60
        cost: 1200
        maxPlaces: 10
        locationId: 5
        name: Лекция по Java
        id: 42
        ownerId: 10
        status: WAIT_START
      properties:
        id:
          example: 42
          format: int64
          type: integer
        name:
          description: Название мероприятия
          example: Лекция по Java
          type: string
        ownerId:
          description: id пользователя-создателя мероприятия
          example: 10
          type: integer
        maxPlaces:
          description: Максимальное кол-во мест на мероприятии
          example: 10
          type: integer
        occupiedPlaces:
          description: Кол-во уже занятых мест (создатель не учитывается при подсчете)
          example: 7
          type: integer
        date:
          description: Дата и время проведения мероприятия. Формат "YYYY-MM-DDThh:mm:ss"
          format: date-time
          type: string
        cost:
          description: Стоимость в рублях
          example: 1200
          minimum: 1
          type: integer
        duration:
          description: Длительность в минутах
          example: 60
          minimum: 30
          type: integer
        locationId:
          description: Идентификатор локации, где проходит мероприятие
          type: integer
        status:
          "$ref": "#/components/schemas/EventStatus"
      required:
      - cost
      - date
      - duration
      - id
      - locationId
      - maxPlaces
      - name
      - occupiedPlaces
      - ownerId
      - status
      type: object
    EventStatus:
      description: |-
        * WAIT_START - мероприятие ожидает начала. Доступные новые регистрации в этом состоянии.
        В этом состоянии еще возможно отменить мероприятие (перевести в статус CANCELLED), в остальных - нельзя.
        * STARTED - мероприятие началось. Новые регистрации недоступны на мероприятие.
        * CANCELLED - мероприятие отменено организтором. Новые регистрации недоступны на мероприятие.
        * FINISHED - мероприятие окончилось. Новые регистрации недоступны на мероприятие.
      enum:
      - WAIT_START
      - STARTED
      - CANCELLED
      - FINISHED
      example: WAIT_START
      type: string
    EventCreateRequestDto:
      description: Объект с данными для создания нового мероприятия.
      example:
        date: '2000-01-23T04:56:07.000+00:00'
        duration: 60
        cost: 1200
        maxPlaces: 10
        locationId: 10
        name: Лекция по Java
      properties:
        name:
          description: Название мероприятия
          example: Лекция по Java
          type: string
        maxPlaces:
          description: Максимальное кол-во мест на мероприятии
          example: 10
          type: integer
        date:
          description: Дата и время проведения мероприятия. Формат "YYYY-MM-DDThh:mm:ss"
          format: date-time
          type: string
        cost:
          description: Стоимость в рублях
          example: 1200
          minimum: 1
          type: integer
        duration:
          description: Длительность в минутах
          example: 60
          minimum: 30
          type: integer
        locationId:
          description: Идентификатор локации, где проходит мероприятие
          example: 10
          type: integer
      required:
      - cost
      - date
      - duration
      - locationId
      - maxPlaces
      - name
      type: object
    EventUpdateRequestDto:
      description: Объект с данными для обновления мероприятия
      example:
        date: '2000-01-23T04:56:07.000+00:00'
        duration: 60
        cost: 3000
        maxPlaces: 100
        locationId: 50
        name: Новое назание
      properties:
        name:
          description: Новое название мероприятия
          example: Новое назание
          type: string
        maxPlaces:
          description: Новое кол-во мест мероприятия. Локация должна позволять вместить
            столько людей. Должно быть больше, чем уже записанных людей.
          example: 100
          type: integer
        date:
          description: Новая дата и время проведения мероприятия. Формат "YYYY-MM-DDThh:mm:ss"
          format: date-time
          type: string
        cost:
          description: Новая стоимость мероприятия в рублях
          example: 3000
          minimum: 1
          type: integer
        duration:
          description: Новая длительность в минутах
          example: 60
          minimum: 30
          type: integer
        locationId:
          description: Новая локация для мероприятия. Локация должна вмещать количество
            пользователей, которые уже записаны. Также локация должна вмещать maxPlaces
            - кол-во мест на мероприятии.
          example: 50
          type: integer
      type: object
    EventSearchRequestDto:
      description: Фильтр для поиска мероприятий. Обязательных полей нет. Если все
        поля не заданы, то должен вернуться список всех мероприятий
      example:
        durationMax: 120
        dateStartBefore: '2000-01-23T04:56:07.000+00:00'
        placesMin: 10
        locationId: 5
        eventStatus: WAIT_START
        name: Лекция
        placesMax: 100
        costMin: 1000
        dateStartAfter: '2000-01-23T04:56:07.000+00:00'
        costMax: 5000
        durationMin: 60
      properties:
        name:
          description: Имя мероприятия для поиска по точному совпадению
          example: Лекция
          type: string
        placesMin:
          description: Минимальная вместимость мероприятия
          example: 10
          type: integer
        placesMax:
          description: Максимальная вместимость мероприятия
          example: 100
          type: integer
        dateStartAfter:
          description: Минимальная дата и время начала мероприятия. Формат "YYYY-MM-DDThh:mm:ss"
          format: date-time
          type: string
        dateStartBefore:
          description: Максимальная дата и время начала мероприятия. Формат "YYYY-MM-DDThh:mm:ss"
          format: date-time
          type: string
        costMin:
          description: Минимальная стоимость участия в рублях
          example: 1000
          type: integer
        costMax:
          description: Максимальная стоимость участия в рублях
          example: 5000
          type: integer
        durationMin:
          description: Минимальная длительность мероприятия (в минутах)
          example: 60
          type: integer
        durationMax:
          description: Максимальная длительность мероприятия (в минутах)
          example: 120
          type: integer
        locationId:
          description: Идентификатор локации мероприятия для фильтрации
          example: 5
          type: integer
        eventStatus:
          "$ref": "#/components/schemas/EventStatus"
      type: object
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http